ipfixprobe - introduce input NFB plugin
