+++ CMakeLists.txt
@@ -14,6 +14,7 @@ set(CMAKE_MODULE_PATH ${CMAKE_SOURCE_DIR}/cmake/modules)
 
 option(ENABLE_INPUT_PCAP "Enable build of input PCAP plugin" ON)
 option(ENABLE_INPUT_DPDK "Enable build of input DPDK plugin" OFF)
+option(ENABLE_INPUT_NFB  "Enable build of input NFB plugin"  OFF)
 
 set(CMAKE_C_STANDARD 11)
 set(CMAKE_C_STANDARD_REQUIRED ON)
