+++ src/plugins/input/CMakeLists.txt
@@ -7,3 +7,7 @@ endif()
 if (ENABLE_INPUT_DPDK)
 	add_subdirectory(dpdk)
 endif()
+
+if (ENABLE_INPUT_NFB)
+	add_subdirectory(nfb)
+endif()
