+++ src/plugins/input/nfb/src/ndpReader.hpp
@@ -1,14 +1,36 @@
-#ifndef NFBREADER_HPP
-#define NFBREADER_HPP
+#pragma once
 
-#include "ndpheader.h"
+#include "ndpHeader.hpp"
 
 #include <string>
 #include <vector>
 
-#include <nfb/ndp.h>
+#include <numa.h>
 #include <stdint.h>
 #include <sys/time.h>
+#include <unistd.h>
+
+extern "C" {
+#include <nfb/ndp.h>
+#include <nfb/nfb.h>
+}
+
+namespace ipxp {
+
+struct NdpReaderContext {
+	void* reader;
+};
+
+extern void ndp_reader_init(struct NdpReaderContext* context);
+extern void ndp_reader_free(struct NdpReaderContext* context);
+extern const char* ndp_reader_error_msg(struct NdpReaderContext* context);
+extern int ndp_reader_init_interface(struct NdpReaderContext* context, const char* interface);
+extern void ndp_reader_print_stats(struct NdpReaderContext* context);
+extern void ndp_reader_close(struct NdpReaderContext* context);
+extern int ndp_reader_get_pkt(
+	struct NdpReaderContext* context,
+	struct ndp_packet** ndp_packet,
+	struct ndp_header** ndp_header);
 
 enum class NdpFwType {
 	NDP_FW_HANIC,
@@ -47,4 +69,4 @@ private:
 	bool ndp_packet_buffer_valid;
 };
 
-#endif // NFBREADER_HPP
+} // namespace ipxp
