--- src/plugins/input/nfbCInterface/configure.ac
@@ -1,11 +1,0 @@
-AC_PREREQ([2.68])
-
-AC_INIT([Nfb Reader Interface], 1.0)
-AM_INIT_AUTOMAKE
-LT_INIT
-
-AC_PROG_CXX
-AC_PROG_CC
-
-AC_CONFIG_FILES(Makefile)
-AC_OUTPUT
